Sweep code for studying model population stats (1 of 2) (#143)
Summary:
This is a *major update* and introduces powerful new functionality to pycls.

The pycls codebase now provides powerful support for studying *design spaces* and more generally *population statistics* of models as introduced in [On Network Design Spaces for Visual Recognition](https://arxiv.org/abs/1905.13214) and [Designing Network Design Spaces](https://arxiv.org/abs/2003.13678). This idea is that instead of planning a single pycls job (e.g., testing a specific model configuration), one can study the behavior of an entire population of models. This allows for quite powerful and succinct experimental design, and elevates the study of individual model behavior to the study of the behavior of model populations. Please see [`SWEEP_INFO`](docs/SWEEP_INFO.md) for details.

This is commit 1 of 2 for the sweep code. It is focused on the sweep config, setting up the sweep, and launching it.

Pull Request resolved: https://github.com/facebookresearch/pycls/pull/143

Reviewed By: pdollar

Differential Revision: D28580825

Pulled By: rajprateek

fbshipit-source-id: 9221f0a9b3651642d2c6acd87befacd6521825cc

Co-authored-by: Raj Prateek Kosaraju <rajprateek@users.noreply.github.com>
Co-authored-by: Piotr Dollar <699682+pdollar@users.noreply.github.com>