Sweep code for studying model population stats (2 of 2) (#144)
Summary:
This is a *major update* and introduces powerful new functionality to pycls.

The pycls codebase now provides powerful support for studying *design spaces* and more generally *population statistics* of models as introduced in [On Network Design Spaces for Visual Recognition](https://arxiv.org/abs/1905.13214) and [Designing Network Design Spaces](https://arxiv.org/abs/2003.13678). This idea is that instead of planning a single pycls job (e.g., testing a specific model configuration), one can study the behavior of an entire population of models. This allows for quite powerful and succinct experimental design, and elevates the study of individual model behavior to the study of the behavior of model populations. Please see [`SWEEP_INFO`](docs/SWEEP_INFO.md) for details.

This is commit 2 of 2 for the sweep code. It is focused on sweep analysis, sweep examples, and documentation.

Pull Request resolved: https://github.com/facebookresearch/pycls/pull/144

Reviewed By: rajprateek

Differential Revision: D28586390

Pulled By: pdollar

fbshipit-source-id: 55856f9aaf7ae49243f4870c787a144b03e5d2a9

Co-authored-by: Raj Prateek Kosaraju <rajprateek@users.noreply.github.com>
Co-authored-by: Piotr Dollar <699682+pdollar@users.noreply.github.com>